requirements.txt
google-cloud-aiplatform>=1.60.0
google-generativeai>=0.8.0
google-cloud-texttospeech>=2.16.0
google-auth>=2.29.0
requests>=2.31.0
python-dotenv>=1.0.0
moviepy>=1.0.3
Pillow>=10.3.0
